S13P31A703-185 be 챌린지 진행 중 실시간 운동 데이터 수집 랭킹 반영 서버 간 동기화 기능 구현

See merge request s13-final/S13P31A703!92

+++ backend/runnity-websocket/src/main/java/com/runnity/websocket/dto/websocket/server/ConnectedMessage.java
@@ -62,7 +62,7 @@ public record ConnectedMessage(
         String nickname,
         String profileImage,
         Double distance,
-        Double pace
+        Integer pace
     ) {
         public Participant {
             if (userId == null) {
